Fixed error prone constructors and renamed, grouped models
