+++ src/RevolutAPI/RevolutAPI/Models/Authorization/RefreshAccessTokenModel.cs
@@ -1,8 +1,4 @@
-﻿using System;
-using System.Collections.Generic;
-using System.Text;
-
-namespace RevolutAPI.Models.Authorization
+﻿namespace RevolutAPI.Models.Authorization
 {
     public class RefreshAccessTokenModel
     {
